Adding SQL and keyvault connections

+++ src/Users/Users.Api/Controllers/V1/UserController.cs
@@ -6,6 +6,10 @@
     using ForexMiner.Heimdallr.Common.Data.Contracts.User;
     using ForexMiner.Heimdallr.Users.Api.Services;
     using Microsoft.AspNetCore.Authorization;
+    using System.Linq;
+    using System.Security.Claims;
+    using ForexMiner.Heimdallr.Common.Data.Exceptions;
+    using System.Net;
 
     [ApiController]
     [ApiVersion("1")]
@@ -37,12 +41,40 @@
         [HttpGet("{userId}")]
         public User GetUser(Guid userId)
         {
+            // Claim validation
+            var userIdClaim = User.Claims.Where(c => c.Type == ClaimTypes.Name).FirstOrDefault();
+            if (userIdClaim == null)
+            {
+                // User somehow doesn't have a user id in the claims
+                throw new ProblemDetailsException(HttpStatusCode.Unauthorized, "Unauthorized access.");
+            }
+            if (Guid.Parse(userIdClaim.Value) != userId)
+            {
+                // User is trying to reach a different user's details
+                throw new ProblemDetailsException(HttpStatusCode.Forbidden, "You don't have permissions to see other user's details.");
+            }
+
+            // Get user from service if validations are passed
             return _userService.GetUserById(userId);
         }
 
         [HttpPatch("{userId}")]
         public User UpdateUser(Guid userId, [FromBody] UserUpdate userUpdate)
         {
+            // Claim validation
+            var userIdClaim = User.Claims.Where(c => c.Type == ClaimTypes.Name).FirstOrDefault();
+            if (userIdClaim == null)
+            {
+                // User somehow doesn't have a user id in the claims
+                throw new ProblemDetailsException(HttpStatusCode.Unauthorized, "Unauthorized access.");
+            }
+            if (Guid.Parse(userIdClaim.Value) != userId)
+            {
+                // User is trying to reach a different user's details
+                throw new ProblemDetailsException(HttpStatusCode.Forbidden, "You don't have permissions to see other user's details.");
+            }
+
+            // Update user if validations are passed
             return _userService.UpdateUser(userId, userUpdate);
         }
 
@@ -57,6 +89,20 @@
         [HttpDelete("{userId}")]
         public void DeleteUser(Guid userId)
         {
+            // Claim validation
+            var userIdClaim = User.Claims.Where(c => c.Type == ClaimTypes.Name).FirstOrDefault();
+            if (userIdClaim == null)
+            {
+                // User somehow doesn't have a user id in the claims
+                throw new ProblemDetailsException(HttpStatusCode.Unauthorized, "Unauthorized access.");
+            }
+            if (Guid.Parse(userIdClaim.Value) != userId)
+            {
+                // User is trying to reach a different user's details
+                throw new ProblemDetailsException(HttpStatusCode.Forbidden, "You don't have permissions to see other user's details.");
+            }
+
+            // Delete user if validations are passed
             _userService.DeleteUser(userId);
         }
     }
